Calculate life pct regularly from Update instead of through Harmony (faster). Started creating graph

+++ PerformanceMeter/PerformanceMeter/Utils.cs
@@ -13,9 +13,9 @@ namespace PerformanceMeter
         /// </summary>
         /// <param name="lifePctFrames">(timeMs,  pct 0.0 to 1.0)</param>
         /// <returns>0 if less than two points, else the average life over time</returns>
-        public static float CalculateAverageLifePercent(Dictionary<float, float> lifePctFrames)
+        public static float CalculateAverageLifePercent(Dictionary<int, float> lifePctFrames)
         {
-            List<float> times = lifePctFrames.Keys.ToList();
+            List<int> times = lifePctFrames.Keys.ToList();
             times.Sort();
 
             if (times.Count < 2)
